requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
anywidget==0.9.18
appnope @ file:///Users/ec2-user/ci_py311/appnope_1678317440516/work
APScheduler==3.11.0
asttokens @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_15z9et5x6m/croot/asttokens_1743630451525/work
asyncua==1.1.6
attrs==25.3.0
av==12.3.0
backoff==2.2.1
bbox_visualizer==0.2.1
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_aef2nlxc2o/croot/comm_1709322866518/work
commonmark==0.9.1
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
debugpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a5sm9dmihj/croot/debugpy_1736267550163/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
efficientnet_pytorch==0.7.1
einops==0.8.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
ezdxf==1.4.1
fairscale==0.4.13
fastapi==0.110.3
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2025.3.2
ftfy==6.3.1
google-crc32c==1.7.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
imageio==2.37.0
inference==0.46.5
inference-sdk==0.46.1
iniconfig==2.1.0
ipykernel @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_6757k8pgf8/croot/ipykernel_1737662465787/work
ipython @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_f6ex7m1blz/croot/ipython_1734550232322/work
ipywidgets==8.1.6
jedi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b78vrox9dk/croot/jedi_1733987406492/work
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jupyter_bbox_widget==0.6.0
jupyter_client @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_e4fuoqieve/croot/jupyter_client_1737570976188/work
jupyter_core @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_80l8bw91xp/croot/jupyter_core_1718818299834/work
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lazy_loader==0.4
logger==1.4
manifold3d==3.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline @ file:///Users/ec2-user/ci_py311/matplotlib-inline_1678317544666/work
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
nest-asyncio @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_c44_atq5c1/croot/nest-asyncio_1708532677863/work
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
nvidia-ml-py==12.570.86
onnx==1.17.0
onnx_graphsurgeon==0.5.8
onnxruntime==1.21.0
onnxsim==0.4.36
open_clip_torch==2.32.0
openai==1.75.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_15t4xe1fp0/croot/packaging_1734472125760/work
paho-mqtt==1.6.1
pandas==2.2.3
parso @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ffagjb_lok/croot/parso_1733963312676/work
peft==0.15.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==11.1.0
pillow_heif==0.22.0
platformdirs @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_640usff2p5/croot/platformdirs_1692205656061/work
pluggy==1.5.0
polygraphy==0.49.20
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
prompt-toolkit @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4d8kk9w3ed/croot/prompt-toolkit_1704404354789/work
propcache==0.3.1
protobuf==6.30.2
psutil @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ed123tcwwq/croot/psutil_1736369384437/work
psygnal==0.13.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pybase64==1.0.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.7.1
pydantic_core==2.33.1
pydot==2.0.0
pyee==13.0.0
pyglet==2.1.3
Pygments @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cflayrfqsi/croot/pygments_1684279981084/work
PyJWT==2.10.1
pylabel==0.1.55
pylibsrtp==0.12.0
pylogix==1.0.5
pymodbus==3.8.3
PyOpenGL==3.1.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyrender==0.1.45
pytest==8.3.5
python-dateutil @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_efk5_uakg8/croot/python-dateutil_1716495742183/work
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_728n2oj6m9/croot/pyzmq_1734709561835/work
redis==5.0.8
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rf100vl==1.0.0
rfdetr==1.1.0
rich==13.0.1
roboflow==1.1.60
rtree==1.4.0
s3transfer==0.10.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a2c0voe1m0/croot/tornado_1733960496093/work
tqdm==4.67.1
traitlets @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_f49losghmg/croot/traitlets_1718227069025/work
transformers==4.51.3
triangle==20230923
trimesh==4.6.6
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_6dn2s8ln8g/croot/typing_extensions_1734714858107/work
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.2
wait_for2==0.3.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==4.0.14
yarl==1.19.0
zxing-cpp==2.2.0
